03_Source_Code/branch_operations/requirements.txt
# Wekeza DFS Branch Operations System - Python Dependencies

# Core Framework
streamlit==1.28.0

# Database Connectivity
mysql-connector-python==8.2.0

# Data Processing
pandas==2.1.0
numpy==1.24.0

# Date/Time Handling
python-dateutil==2.8.2

# Image Processing (for file uploads)
Pillow==10.0.0

# Data Visualization (for reports)
plotly==5.17.0

# Additional Utilities
uuid==1.30
datetime

# Development Dependencies (optional)
pytest==7.4.0
pytest-cov==4.1.0

# Security (for production)
cryptography==41.0.0

# Environment Management
python-dotenv==1.0.0